Refactor `DSTypography`, add screen `DSLegacyTypography`
