+++ ts/features/design-system/core/DSColors.tsx
@@ -1,6 +1,5 @@
 import {
   H3,
-  H4,
   H6,
   IOColorGradients,
   IOColors,
