+++ ts/features/design-system/navigation/navigator.tsx
@@ -73,6 +73,7 @@ import { DSTypography } from "../core/DSTypography";
 import { DSWallet } from "../core/DSWallet";
 import { DSWizardScreen } from "../core/DSWizardScreen";
 import { DSLegacyAdvice } from "../core/DSLegacyAdvice";
+import { DSLegacyTypography } from "../core/DSLegacyTypography";
 import { DesignSystemParamsList } from "./params";
 import DESIGN_SYSTEM_ROUTES from "./routes";
 
@@ -528,6 +529,14 @@ export const DesignSystemNavigator = () => {
         </Stack.Group>
 
         {/* LEGACY */}
+        <Stack.Screen
+          name={DESIGN_SYSTEM_ROUTES.LEGACY.TYPOGRAPHY.route}
+          component={DSLegacyTypography}
+          options={{
+            headerTitle: DESIGN_SYSTEM_ROUTES.LEGACY.TYPOGRAPHY.title
+          }}
+        />
+
         <Stack.Screen
           name={DESIGN_SYSTEM_ROUTES.LEGACY.PICTOGRAMS.route}
           component={DSLegacyPictograms}
