+++ ts/features/design-system/navigation/params.ts
@@ -45,6 +45,7 @@ export type DesignSystemParamsList = {
   [DESIGN_SYSTEM_ROUTES.DEBUG.FOOTER_ACTIONS_STICKY.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.DEBUG.FOOTER_ACTIONS_NOT_FIXED.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.DEBUG.SCREEN_END_MARGIN.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.TYPOGRAPHY.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.LEGACY.PICTOGRAMS.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.LEGACY.BUTTONS.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.LEGACY.TEXT_FIELDS.route]: undefined;
